+++ 권석원/[6주차]DP/8-7.py
@@ -1,0 +1,13 @@
+import sys
+
+n = int(sys.stdin.readline())
+
+d = [0] * 1001
+
+d[1] = 1
+d[2] = 3
+
+for i in range(3,n+1):
+    d[i] = d[i-1] + (d[i*2] * 2) % 796796
+
+print(d[n])
